% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirallele.R
\docType{methods}
\name{getAllelesFromPython,PythonDB-method}
\alias{getAllelesFromPython,PythonDB-method}
\title{Get list of \code{KIRAllele}s from a \code{PythonDB} object.}
\usage{
\S4method{getAllelesFromPython}{PythonDB}(db, ncores = parallel::detectCores()
  - 4)
}
\arguments{
\item{{db}{A \code{PythonDB object from which we want to extract the allels}}}

\item{{nccores}{The number of CPU cores that should be assigned this job. It is not huge computations carried out by this function,
but parallel makes it a little bit faster anyway.}}
}
\value{
{A \code{list} of \code{KIRAllele} objects}
}
\description{
Get list of \code{KIRAllele}s from a \code{PythonDB} object.
}

