% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirallele.R
\docType{class}
\name{KIRAllele-class}
\alias{KIRAllele-class}
\title{Class \code{KIRAllele}}
\description{
A wrapper extending \linkS4class{HLAAllele} imported from \link{hlatools}
No methods or attributes in addition to the ones inherrited from \linkS4class{HLAAllele} are implemented in this implementation.
This class should allow to handle KIR alleles in a separate object.
}

