% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirallele.R
\name{KIRAlleleFromDB}
\alias{KIRAlleleFromDB}
\title{Extract alleles for specific KIR gene}
\usage{
KIRAlleleFromDB(genename = "KIR2DL1")
}
\arguments{
\item{genename}{The name of the KIR gene that should be pulled. Here it defaults to KIR2DL1.}
}
\value{
A \linkS4class{KIRAllele} object.
}
\description{
This function pulls all alleles from the data base file that are linked to one particular KIR gene.
The alleles are returned as \linkS4class{KIRAllele} object
}

