% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirallele.R
\docType{class}
\name{KIRAllele-class}
\alias{KIRAllele-class}
\title{Class \code{KIRAllele}}
\description{
{A wrapper extending \code{HLAAllele-class} imported from \code{hlatools}.
No methods or attributes in addition to the ones inherrited from \code{HLAAllele} are implemented in this implementation.
This class should allows to handle KIR alleles in a separate object.}
}

