% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirallele.R
\docType{class}
\name{PythonDB-class}
\alias{PythonDB-class}
\title{Class \code{PythonDB}}
\description{
{S4 representing a KIR data base parsed from a Genebank file using a Python script.}
}
\section{Slots}{

\describe{
\item{\code{{pydb}}{A \code{list_Python} object that hold the raw data obtained from a Genebank file.}}

\item{\code{{locusNames}}{A \code{character} vector with the KIR gene name of each allele in \code{pydb}}}

\item{\code{{dbPath}}{A \code{character} with a reference to the Genebank data file used to initialize the object}}

\item{\code{{locusType}}{\code{character} that defaults to 'KIR'. The python script could also handle locuses other than KIR.
However we do not use this feature in this package.}}
}}

